refactor(sentry): split the initializer into focused modules
